Move stuff around and condition
