+++ airbyte-integrations/connectors/destination-bigquery/src/main/kotlin/io/airbyte/integrations/destination/bigquery/spec/BigqueryConfiguration.kt
@@ -7,20 +7,9 @@ package io.airbyte.integrations.destination.bigquery.spec
 import io.airbyte.cdk.load.command.DestinationConfiguration
 import io.airbyte.cdk.load.command.DestinationConfigurationFactory
 import io.airbyte.cdk.load.command.gcs.GcsClientConfiguration
-import io.airbyte.cdk.load.command.object_storage.ObjectStorageCompressionConfiguration
-import io.airbyte.cdk.load.command.object_storage.ObjectStorageCompressionConfigurationProvider
-import io.airbyte.cdk.load.command.object_storage.ObjectStoragePathConfiguration
-import io.airbyte.cdk.load.command.object_storage.ObjectStoragePathConfigurationProvider
-import io.airbyte.cdk.load.command.s3.S3ClientConfiguration
-import io.airbyte.cdk.load.command.s3.S3ClientConfigurationProvider
-import io.airbyte.cdk.load.file.NoopProcessor
 import io.airbyte.cdk.load.write.db.DbConstants
 import io.micronaut.context.annotation.Factory
-import io.micronaut.context.annotation.Requires
-import io.micronaut.context.condition.Condition
-import io.micronaut.context.condition.ConditionContext
 import jakarta.inject.Singleton
-import java.io.ByteArrayOutputStream
 
 data class BigqueryConfiguration(
     val projectId: String,
@@ -36,6 +25,8 @@ data class BigqueryConfiguration(
 sealed interface LoadingMethodConfiguration
 
 data object BatchedStandardInsertConfiguration : LoadingMethodConfiguration
+
+data object BulkLoadConfiguration : LoadingMethodConfiguration
 
 data class GcsStagingConfiguration(
     val gcsClientConfig: GcsClientConfiguration,
@@ -74,32 +65,4 @@ class BigqueryConfigurationFactory :
 @Factory
 class BigqueryConfigurationProvider(private val config: DestinationConfiguration) {
     @Singleton fun get() = config as BigqueryConfiguration
-}
-
-class BigQueryIsConfiguredForBulkLoad : Condition {
-    override fun matches(context: ConditionContext<*>): Boolean {
-        val config = context.beanContext.getBean(BigqueryConfiguration::class.java)
-        return config.bigqueryLoadTypeConfiguration.loadTypeConfiguration is BulkLoadConfiguration
-    }
-}
-
-@Singleton
-@Requires(condition = BigQueryIsConfiguredForBulkLoad::class)
-class BigQueryBulkLoadConfiguration(
-    private val config: BigqueryConfiguration,
-) :
-    ObjectStoragePathConfigurationProvider,
-    ObjectStorageCompressionConfigurationProvider<ByteArrayOutputStream>,
-    S3ClientConfigurationProvider {
-    override val s3ClientConfiguration: S3ClientConfiguration
-        get() = TODO("Not yet implemented")
-    override val objectStoragePathConfiguration =
-        ObjectStoragePathConfiguration(
-            prefix = "",
-            pathPattern = "\${NAMESPACE}/\${STREAM_NAME}/\${YEAR}/\${MONTH}/\${DAY}/\${EPOCH}/",
-            fileNamePattern = "{part_number}{format_extension}",
-        )
-    override val objectStorageCompressionConfiguration:
-        ObjectStorageCompressionConfiguration<ByteArrayOutputStream> =
-        ObjectStorageCompressionConfiguration(NoopProcessor)
 }
